requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.2.5
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.6
zope.interface==6.0
